Update patch set 1
Patch Set 1:

Okay man :) And a quick question, do you have an idea how to fix the powersave governor problem? Instead of running in the minimum frequency (250Mhz) its stuck on 600Mhz and not going below that, as it should. Can you help with that problem? I'm out of ideas how to solve it :(

Patch-set: 1