Update patch set 2
Patch Set 2:

Can you talk to Ethan? He can easily push it over there... I'm afraid I will mess it up, because the last time Ethan was the one who pushed it... I messed up some files. Just talk to him, because he wants to know the origin and the firmware of the MK4 update, as it doesn't show in the Samsung release site

Patch-set: 2