Update patch set 1
Patch Set 1:

Part of that is my fault, I haven't had time to keep up with the sheer volume of patches coming in, i9500 stuff tends to be overlook by me a lot. The RIL needs general updating though, not just flipping the enums, and as was mentioned before, this isn't a real fix. I probably need to revisit a lot of i9500 stuff in general...

Patch-set: 1