Update patch set 1
Patch Set 1:

And if not, I will just focus on developing for the Note 4 that I'm about to receive. Will see if the people there are more open to important fixes and getting to a ROM that's actually enjoyable and usuable for the users out there. And if not, I will just say this whole Cyanogenmod project goodbye. I mean what's the use of fixing stuff when none of your fixes get merged ? I can't think of a bigger waste of time ...

Patch-set: 1