Update patch set 4
Patch Set 4:

If you remove the mp3 library you ought to remove the ALP audio flag as well. Give this a try, it should do the right thing for nuplayer. AwesomePlayer has a hardcoded entry for the MP3 decoder regardless of media_codecs, so that may still not work.

Patch-set: 4